chore: extract function into own file
